feat: Convert component to styleguide component
