+++ packages/design/src/index.scss
@@ -10,6 +10,7 @@
 @forward 'components/definition-list.scss';
 @forward 'components/headings.scss';
 @forward 'components/layer.scss';
+@forward 'components/metrics-wheel.scss';
 @forward 'components/tag.scss';
 @forward 'components/tile.scss';
 @forward 'components/tooltip.scss';
