+++ packages/react/src/main.tsx
@@ -6,6 +6,7 @@ export { CodeSnippet } from './components/code-snippet.tsx'
 export { DefinitionList, DefinitionTerm, DefinitionDescription } from './components/definition-list.tsx'
 export type { DefinitionListProps } from './components/definition-list.tsx'
 export { Layer } from './components/layer.tsx'
+export { MetricsWheel } from './components/metrics-wheel.tsx'
 export { Tile } from './components/tile.tsx'
 export { Tooltip, TooltipProvider } from './components/tooltip.tsx'
 export { Tag } from './components/tag.tsx'
